ai_engine/requirements.txt
Flask
flask-cors
sentence-transformers
faiss-cpu
numpy
pdfplumber
requests
